Update task.controller.ts and task.service.ts

+++ src/task/task.service.ts
@@ -1,9 +1,16 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  HttpCode,
+  HttpStatus,
+  Injectable,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from 'src/entities/task.entity';
-import { formatDate } from 'src/helpers/date-and-time';
+import { formatDate } from 'src/helpers/date-and-time.helper';
 import { Repository } from 'typeorm';
 import { UpdateTaskDto } from './dto/update-task.dto';
+import { GetUser } from 'src/decorator/get-user.decorator';
+import { User } from 'src/entities/user.entity';
 
 @Injectable()
 export class TaskService {
@@ -46,9 +53,12 @@ export class TaskService {
     return await this.taskRepository.update(userId, dto);
   }
 
-  async deleteTask(userId: number): Promise<void> {
-    const task = await this.taskRepository.findOneBy({ id: userId });
-
+  @HttpCode(204)
+  async deleteTask(id: number, @GetUser() user: User): Promise<void> {
+    const task = await this.taskRepository.findOneBy({ id, user });
+    if (!task) {
+      throw new BadRequestException('Task not found!');
+    }
     await this.taskRepository.delete(task);
   }
 }
